+++ tests/testthat/test-add_n.R
@@ -147,6 +147,7 @@ test_that("tidy_add_n() works with variables having non standard name", {
 
 
 test_that("tidy_add_n() works with lme4::lmer", {
+  skip_on_cran()
   df <- gtsummary::trial
   df$stage <- as.character(df$stage)
   df$group <- rep.int(1:2, 100)
@@ -156,6 +157,7 @@ test_that("tidy_add_n() works with lme4::lmer", {
 
 
 test_that("tidy_add_n() works with lme4::glmer", {
+  skip_on_cran()
   df <- gtsummary::trial
   df$stage <- as.character(df$stage)
   df$group <- rep.int(1:2, 100)
@@ -202,6 +204,14 @@ test_that("tidy_add_n() works with nnet::multinom", {
     res$n_event,
     c(57, 21, 16, 8, 57, 57, 58, 12, 18, 12, 58, 58)
   )
+
+  # when y is not coded as a factor
+  mod <- nnet::multinom(race ~ age + lwt + bwt, data = MASS::birthwt)
+  expect_error(
+    mod %>% tidy_and_attach() %>% tidy_add_n(),
+    NA
+  )
+
 })
 
 test_that("tidy_add_n() works with survey::svyglm", {