+++ tests/testthat/test-model_get_n.R
@@ -188,6 +188,7 @@ test_that("model_get_n() works with stats::poly()", {
 
 
 test_that("model_get_n() works with lme4::lmer", {
+  skip_on_cran()
   df <- gtsummary::trial
   df$stage <- as.character(df$stage)
   df$group <- rep.int(1:2, 100)
@@ -198,6 +199,7 @@ test_that("model_get_n() works with lme4::lmer", {
 
 
 test_that("model_get_n() works with lme4::glmer", {
+  skip_on_cran()
   df <- gtsummary::trial
   df$stage <- as.character(df$stage)
   df$group <- rep.int(1:2, 100)
@@ -256,6 +258,10 @@ test_that("model_get_n() works with nnet::multinom", {
     res$n_event,
     c(57, 16, 8, 12, 57, 57, 21, 58, 18, 12, 16, 58, 58, 12)
   )
+
+  # when y is not coded as a factor
+  mod <- nnet::multinom(race ~ age + lwt + bwt, data = MASS::birthwt)
+  expect_true(mod %>% model_get_n() %>% nrow() > 0)
 })
 
 test_that("model_get_n() works with survey::svyglm", {